+++ src/clj/skylobby/fx/battle.clj
@@ -13,6 +13,7 @@
     [skylobby.fx.mod-sync :refer [mod-sync-pane]]
     [skylobby.fx.mods :refer [mods-view]]
     [skylobby.fx.players-table :refer [players-table]]
+    [skylobby.resource :as resource]
     [spring-lobby.fx.font-icon :as font-icon]
     [spring-lobby.fs :as fs]
     [spring-lobby.spring :as spring]
@@ -38,14 +39,14 @@
   [:archiving :auto-get-resources :battles :battle :battle-players-color-allyteam :bot-name
    :bot-username :bot-version :channels :chat-auto-scroll :cleaning :client-data :copying :downloadables-by-url :drag-allyteam :drag-team :engine-filter :engine-version
    :engines :extracting :file-cache :git-clone :gitting :http-download :importables-by-path
-   :isolation-type :map-filter
+   :isolation-type
    :map-input-prefix :map-details :maps :message-drafts :minimap-type :mod-details :mod-filter :mods :parsed-replays-by-path :rapid-data-by-id :rapid-data-by-version
    :rapid-download :rapid-update :server-key :spring-isolation-dir :spring-settings :springfiles-search-results :tasks-by-type :update-engines :update-maps :update-mods :username :users])
 
 (defn battle-view
   [{:keys [auto-get-resources battle battles battle-players-color-allyteam bot-name bot-username bot-version
            channels chat-auto-scroll client-data downloadables-by-url
-           drag-allyteam drag-team engine-filter engines file-cache http-download map-filter
+           drag-allyteam drag-team engine-filter engines file-cache http-download
            map-input-prefix map-details maps message-drafts minimap-type mod-details mod-filter mods parsed-replays-by-path rapid-data-by-id rapid-data-by-version rapid-download server-key
            spring-isolation-dir spring-settings tasks-by-type users username]
     :as state}]
@@ -66,8 +67,10 @@
         battle-details (spring/battle-details {:battle battle :battles battles :users users})
         engine-version (:battle-version battle-details)
         engine-details (spring/engine-details engines engine-version)
-        battle-map-details (get map-details battle-map)
-        battle-mod-details (get mod-details battle-modname)
+        indexed-map (->> maps (filter (comp #{battle-map} :map-name)) first)
+        battle-map-details (get map-details (resource/details-cache-key indexed-map))
+        indexed-mod (->> mods (filter (comp #{battle-modname} :mod-name)) first)
+        battle-mod-details (get mod-details (resource/details-cache-key indexed-mod))
         in-sync (boolean (and (seq battle-map-details)
                               (seq battle-mod-details)
                               (seq engine-details)))
@@ -259,8 +262,7 @@
                       :engine-version engine-version
                       :engines engines
                       :tasks-by-type tasks-by-type
-                      :on-value-changed {:event/type :spring-lobby/assoc-in
-                                         :path [:by-server :local :battles :singleplayer :battle-version]}
+                      :on-value-changed {:event/type :spring-lobby/singleplayer-engine-changed}
                       :spring-isolation-dir spring-isolation-dir}
                      (if (seq engine-details)
                        {:fx/type mods-view
@@ -273,20 +275,18 @@
                         :rapid-data-by-id rapid-data-by-id
                         :rapid-download rapid-download
                         :tasks-by-type tasks-by-type
-                        :on-value-changed {:event/type :spring-lobby/assoc-in
-                                           :path [:by-server :local :battles :singleplayer :battle-modname]}
+                        :on-value-changed {:event/type :spring-lobby/singleplayer-mod-changed}
                         :spring-isolation-dir spring-isolation-dir}
                        {:fx/type :label
                         :text " Game: Get an engine first"})
                      {:fx/type maps-view
                       :downloadables-by-url downloadables-by-url
                       :http-downloads http-download
-                      :map-filter map-filter
+                      :map-input-prefix map-input-prefix
                       :map-name battle-map
                       :maps maps
                       :tasks-by-type tasks-by-type
-                      :on-value-changed {:event/type :spring-lobby/assoc-in
-                                         :path [:by-server :local :battles :singleplayer :battle-map]}
+                      :on-value-changed {:event/type :spring-lobby/singleplayer-map-changed}
                       :spring-isolation-dir spring-isolation-dir}]}
                    {:fx/type :flow-pane
                     :vgap 5
@@ -326,7 +326,7 @@
                           :alignment :center-left
                           :children
                           [{:fx/type maps-view
-                            :map-filter map-filter
+                            :map-input-prefix map-input-prefix
                             :map-name battle-map
                             :maps maps
                             :spring-isolation-dir spring-isolation-dir
@@ -348,6 +348,7 @@
                           :battle-mod-details battle-mod-details
                           :engine-details engine-details
                           :engine-file engine-file
+                          :indexed-mod indexed-mod
                           :mod-details-tasks mod-details-tasks
                           :mod-update-tasks mod-update-tasks
                           :rapid-tasks-by-id rapid-tasks-by-id}
@@ -356,6 +357,7 @@
                          {:fx/type map-sync-pane
                           :battle-map battle-map
                           :battle-map-details battle-map-details
+                          :indexed-map indexed-map
                           :import-tasks import-tasks
                           :map-update-tasks map-update-tasks}
                          (select-keys state [:copying :downloadables-by-url :file-cache :http-download :importables-by-path :maps :spring-isolation-dir :tasks-by-type :update-maps]))])})])}}